+++ tools4j-sequences/src/main/java/com/github/panpf/tools4j/sequences/ConstrainedOnceSequence.java
@@ -23,9 +23,10 @@ import java.util.concurrent.atomic.AtomicReference;
 
 class ConstrainedOnceSequence<T> implements Sequence<T> {
 
+    @NotNull
     private final AtomicReference<Sequence<T>> sequenceRef;
 
-    ConstrainedOnceSequence(Sequence<T> sequence) {
+    ConstrainedOnceSequence(@NotNull Sequence<T> sequence) {
         this.sequenceRef = new AtomicReference<>(sequence);
     }