+++ tools4j-sequences/src/main/java/com/github/panpf/tools4j/sequences/FlatteningSequence.java
@@ -25,14 +25,14 @@ import java.util.NoSuchElementException;
 
 public class FlatteningSequence<T, R, E> implements Sequence<E> {
 
-    @NotNull
+    @Nullable
     private final Sequence<T> sequence;
     @NotNull
     private final Transformer<T, R> transformer;
     @NotNull
     private final Transformer<R, Iterator<E>> iteratorTransformer;
 
-    public FlatteningSequence(@NotNull Sequence<T> sequence, @NotNull Transformer<T, R> transformer, @NotNull Transformer<R, Iterator<E>> iterator) {
+    public FlatteningSequence(@Nullable Sequence<T> sequence, @NotNull Transformer<T, R> transformer, @NotNull Transformer<R, Iterator<E>> iterator) {
         this.sequence = sequence;
         this.transformer = transformer;
         this.iteratorTransformer = iterator;
@@ -43,8 +43,8 @@ public class FlatteningSequence<T, R, E> implements Sequence<E> {
     public Iterator<E> iterator() {
         return new Iterator<E>() {
 
-            @NotNull
-            private final Iterator<T> iterator = sequence.iterator();
+            @Nullable
+            private final Iterator<T> iterator = sequence != null ? sequence.iterator() : null;
             @Nullable
             private Iterator<E> itemIterator = null;
 
@@ -65,7 +65,7 @@ public class FlatteningSequence<T, R, E> implements Sequence<E> {
                 }
 
                 while (itemIterator == null) {
-                    if (!iterator.hasNext()) {
+                    if (iterator == null || !iterator.hasNext()) {
                         return false;
                     } else {
                         T element = iterator.next();