+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Change:
 Remove:
 * remove: Arrayx remove asList method
 
+Improve:
+* improve: Sequencex.map\* method can now pass in null
+
 ## v1.0.0-beta02
 improve: Improve code
 upgrade: Kotlin upgraded to 1.4.0